change working directory on start
When launching from a context menu, the working directory will be set to
the folder grepWin has been invoked from. With the process now having an
association with this directory, actively attempting to manipulate
(e.g. move/delete) will fail due to an "in use" error. To allow users
to freely manipulate their file systems (without having to close
grepWin), change the working directory to the install path of grepWin.

Signed-off-by: James Knight <james.d.knight@live.com>