correct portable detection
A previous change [1] adjusted some locals to share the module's
directory between the global ini-path and to forward into a new working
directory change (`_wchdir`); however, these changes broken the portable
detection logic by mixing up the usage of `GetModuleDir` and
`GetModulePath`.

This commit restores the use of `GetModulePath` for the portable check.

[1]: 65d06ce3d532dfda880edd764f79b1e43b69c668

Signed-off-by: James Knight <james.d.knight@live.com>