ignore windows system directory as a fallback search-path
Changes have been introduced into grepWin where the current working
directory is used as fallback search path to use when no path is
provided from the command line/context-menu invoke. A possible
undesired effect from this change is that when grepWin is invoked from
a taskbar pin or the start-menu, the default working directory will be
the Windows system path. Users starting grepWin without a possible
target directory are no longer presented with the last saved search
folder.

This commit attempts to allow users to still fallback to the last
searched folder by ignoring the case where the system directory is
detected as the current working directory.

An undesired effect of this change is that if a user was in the system
path and invoked grepWin without a path argument, they would not be
presented an instance where the search path matched their working
directory (note: context-menu invokes still work as expected). While
this may be undesired, it is assumed that the probability of that use
case is significantly smaller than a user invoking grepWin from a pinned
taskbar.

Signed-off-by: James Knight <james.d.knight@live.com>